Add files via upload
Problem Statement:
The focus of this research paper is to answer the following question: Does order of entry in an industrial goods industry have a major role to play in sustaining competitive advantages? Empirical evidence in various papers including the same analysis done for the consumer goods industries suggest that there is direct relationship between the order of entry and market share of the pioneers. Companies like Coca-cola, Hallmark and Birds eyes have been market leaders for decades. This research paper caters to whether there is the same effect of this in industrial goods industries.
Factors that can impact the Problem Statement: 
There are two theories that can limit the possible effect of this study. Firstly, Pioneer advantage can be realized only when the pioneer is successful in understanding the consumer preferences by continually improving the product and keeping their differentiating factors intact against its competitors. Secondly, there is a possibility that the late entrants keeps a watch on the mistakes of the pioneers and then take strategic advantage of this by investing much lesser time and money on their R&Ds. This can give a substantial advantage to the late entrants by adopting “imitate and improve” strategy.
In addition to this, adopting trends in the market can also help the pioneers define their strategy to remain in the market. For example, Artificial Intelligence can be a key trend for a music industry today. Companies like Spotify has to keep up with 20,000 new songs that gets uploaded every day and categorize them based on genres and liking of a consumer; and then suggest them recommendations. Companies like Sony Music Entertainment, which has been a pioneer in the industry since the 2000s are struggling to keep up with these digital trends, however, they still manage to stay in the top game when it comes to records and labels. Companies like Spotify and Apple are the market pioneers for this online digital streaming. This particular example can cater to the issue of how different trends can lead to having different pioneers in the industry for those trends.
Validation of the data:
The analysis of this paper includes Robinson and Fornell’s paper validation with industrial markets. Market pioneer in the PIMS (Profit Impact of Market Strategies) data has its findings in line with the consumer goods industries. The data suggest that the pioneers have substantially high market shares than the late entrants. The market share here depends on both business and industry characteristics. 
In the PIMS data, pioneer effect or the order of entry is based on entrance of business in the market. Businesses are classified into three categories: 
-	Pioneers ( Businesses that are the first in developing the product or service)
-	Early Follower (Businesses which are early followers of the pioneers in still growing market)
-	Late Entrant (Businesses that came late in the already established market)

Data and the model: 
The research paper covers the model, hypothesis, data definitions, descriptive statistics, exploratory data analysis, results and findings, summary and conclusions. The appendix and the references are attached at the last of the paper. The model that fitted the most here was multi-stage OLS regression which addresses the case of multicollinearity and endogeneity in the data.